+++ R/qdap-package.R
@@ -453,5 +453,5 @@ NULL
 #' @usage data(hamlet) 
 #' @format A data frame with 2007 rows and 7 variables 
 #' @references 
-#' \url{http://www.gutenberg.org/ebooks/1524}
+#' \url{http://www.gutenberg.org}
 NULL